[wip]: change how event are handled
The event listeners won't be "filters" and modify any behavior, instead
they will be just a notification that something happened.

Each event is scheduled by the event thread and executes in its own
thread, in parallel to other event listeners for the same event, but on
the same executor (so in general events will always be started in the
order they are scheduled, but not necessarily finish in order).

Exception in events will generate an "exception" event for the same
listener and its parent.

As result, emitting an event becomes really fast operation, by not
blocking the event queue.